+++ e2e/integration/frontity-01/slot-and-fill.spec.js
@@ -1,6 +1,6 @@
 describe("useFills", () => {
   it("should show empty slots", () => {
-    cy.visit("http://localhost:3001/?name=slot-and-fill");
+    cy.visit("http://localhost:3001/?frontity_name=slot-and-fill");
     cy.get("#useFills-slot1-empty")
       .should("exist")
       .should("have.text", "useFills slot 1 empty");
@@ -10,7 +10,7 @@ describe("useFills", () => {
   });
 
   it("should show first Fill", () => {
-    cy.visit("http://localhost:3001/?name=slot-and-fill");
+    cy.visit("http://localhost:3001/?frontity_name=slot-and-fill");
     cy.get("#addFill1").click();
     cy.get("#useFills-slot1-empty").should("not.exist");
     cy.get("#Fill1").should("exist").should("have.text", "I am Fill1");
@@ -18,7 +18,7 @@ describe("useFills", () => {
   });
 
   it("should show second Fill", () => {
-    cy.visit("http://localhost:3001/?name=slot-and-fill");
+    cy.visit("http://localhost:3001/?frontity_name=slot-and-fill");
     cy.get("#addFill2").click();
     cy.get("#useFills-slot1-empty").should("not.exist");
     cy.get("#Fill2").should("exist").should("have.text", "I am Fill2");
@@ -26,7 +26,7 @@ describe("useFills", () => {
   });
 
   it("should show both Fills in priority order", () => {
-    cy.visit("http://localhost:3001/?name=slot-and-fill");
+    cy.visit("http://localhost:3001/?frontity_name=slot-and-fill");
     cy.get("#addFill1").click();
     cy.get("#addFill2").click();
     cy.get("#useFills-slot1-empty").should("not.exist");
@@ -35,7 +35,7 @@ describe("useFills", () => {
   });
 
   it("should show the debug fills when `state.frontity.debug` is true", () => {
-    cy.visit("http://localhost:3001/?name=slot-and-fill");
+    cy.visit("http://localhost:3001/?frontity_name=slot-and-fill");
     cy.get("#toggleDebug").click();
     cy.get("#useFills-slot1-empty").should("not.exist");
     cy.get("#useFills-slot2-empty").should("not.exist");