+++ e2e/integration/frontity-01/tiny-router.spec.js
@@ -1,17 +1,17 @@
 describe("Tiny Router", () => {
   beforeEach(() => {
-    cy.visit("http://localhost:3001?name=tiny-router");
+    cy.visit("http://localhost:3001?frontity_name=tiny-router");
   });
 
   it("should show render contents in home", () => {
-    cy.location("href").should("eq", "http://localhost:3001/?name=tiny-router");
+    cy.location("href").should("eq", "http://localhost:3001/");
     cy.get('[data-test-id="content"]')
       .should("exist")
       .should("have.text", "Home");
   });
 
   it("should switch route when using set", () => {
-    cy.location("href").should("eq", "http://localhost:3001/?name=tiny-router");
+    cy.location("href").should("eq", "http://localhost:3001/");
 
     cy.get('button[data-button-id="switch-to-about"]').click();
     cy.location("href").should("eq", "http://localhost:3001/about/");
@@ -20,14 +20,14 @@ describe("Tiny Router", () => {
       .should("have.text", "About");
 
     cy.get('button[data-button-id="switch-to-home"]').click();
-    cy.location("href").should("eq", "http://localhost:3001/?name=tiny-router");
+    cy.location("href").should("eq", "http://localhost:3001/");
     cy.get('[data-test-id="content"]')
       .should("exist")
       .should("have.text", "Home");
   });
 
   it("should switch route when using set with `push`", () => {
-    cy.location("href").should("eq", "http://localhost:3001/?name=tiny-router");
+    cy.location("href").should("eq", "http://localhost:3001/");
 
     cy.get('button[data-button-id="switch-using-push"]').click();
     cy.location("href").should("eq", "http://localhost:3001/about/");
@@ -37,7 +37,7 @@ describe("Tiny Router", () => {
   });
 
   it("should switch route when using set with `replace`", () => {
-    cy.location("href").should("eq", "http://localhost:3001/?name=tiny-router");
+    cy.location("href").should("eq", "http://localhost:3001/");
 
     cy.get('button[data-button-id="switch-using-replace"]').click();
     cy.location("href").should("eq", "http://localhost:3001/about/");
@@ -47,7 +47,7 @@ describe("Tiny Router", () => {
   });
 
   it("should replace history entry when using set with `replace`", () => {
-    cy.location("href").should("eq", "http://localhost:3001/?name=tiny-router");
+    cy.location("href").should("eq", "http://localhost:3001/");
     cy.get('[data-test-id="content"]')
       .should("exist")
       .should("have.text", "Home");
@@ -65,14 +65,14 @@ describe("Tiny Router", () => {
       .should("have.text", "Privacy");
 
     cy.go("back");
-    cy.location("href").should("eq", "http://localhost:3001/?name=tiny-router");
+    cy.location("href").should("eq", "http://localhost:3001/");
     cy.get('[data-test-id="content"]')
       .should("exist")
       .should("have.text", "Home");
   });
 
   it("should recover state when moving back and forward in history", () => {
-    cy.location("href").should("eq", "http://localhost:3001/?name=tiny-router");
+    cy.location("href").should("eq", "http://localhost:3001/");
 
     cy.get('button[data-button-id="switch-using-state"]').click();
     cy.location("href").should("eq", "http://localhost:3001/about/");
@@ -84,7 +84,7 @@ describe("Tiny Router", () => {
       .should("have.text", "Router has state!");
 
     cy.get('button[data-button-id="switch-to-home"]').click();
-    cy.location("href").should("eq", "http://localhost:3001/?name=tiny-router");
+    cy.location("href").should("eq", "http://localhost:3001/");
     cy.get('[data-test-id="content"]')
       .should("exist")
       .should("have.text", "Home");
@@ -100,7 +100,7 @@ describe("Tiny Router", () => {
       .should("have.text", "Router has state!");
 
     cy.go("back");
-    cy.location("href").should("eq", "http://localhost:3001/?name=tiny-router");
+    cy.location("href").should("eq", "http://localhost:3001/");
     cy.get('[data-test-id="content"]')
       .should("exist")
       .should("have.text", "Home");
@@ -117,7 +117,7 @@ describe("Tiny Router", () => {
   });
 
   it("should work if link doesn't have a trailing forward slash", () => {
-    cy.location("href").should("eq", "http://localhost:3001/?name=tiny-router");
+    cy.location("href").should("eq", "http://localhost:3001/");
 
     cy.get('button[data-button-id="switch-to-about-no-trailing"]').click();
     cy.location("href").should("eq", "http://localhost:3001/about/");
@@ -127,13 +127,10 @@ describe("Tiny Router", () => {
   });
 
   it("should scroll to the element when a hash link is clicked", () => {
-    cy.location("href").should("eq", "http://localhost:3001/?name=tiny-router");
+    cy.location("href").should("eq", "http://localhost:3001/");
     cy.get('div[id="hash-element"]').isNotInViewport();
     cy.get('a[data-link-id="hash-link"]').click();
     cy.get('div[id="hash-element"]').isInViewport();
-    cy.location("href").should(
-      "eq",
-      "http://localhost:3001/?name=tiny-router#hash-element"
-    );
+    cy.location("href").should("eq", "http://localhost:3001/#hash-element");
   });
 });